Avoid gl_PointSize in HeadNode
The Qt RHI does not support point size for all APIs, for example Vulkan
does not support this at all and it seems like it isn't even implemented
for OpenGL at the moment.

Thus to improve compatibility, we explicitly draw two triangles to get
thickness.

This should work with all rendering APIs.