Use the proper window handle to generate textures
Previously this was a total hack, as we just created a dummy
QQuickWindow. This doesn't work anymore and even leads to crashes in Qt
6.5 as createTextureFromImage() returns a nullptr.

To fix this, pass the real QQuickWindow all the way down to the Item, so
that the texture can be generated properly without any issues.

This adds a lot of boilerplate code unfortunately, but it's for the
better if this fixes it properly.