Display ping from other players
This required some severe redesign in the network handling of ping and
pong packets.

Now all players can see the ping from all players.

An improvement for the future could be removing the forced refresh of
the entire player model, when we really only need to refresh the ping
value.