Do not move Explosion to parent thread
Otherwise we get a warning like this:

QObject: Cannot create children for a parent that is in a different thread.
(Parent is Curver(0x555658c59b40), parent's thread is QThread(0x555658a991d0), current thread is QSGRenderThread(0x555658bd9810)

Note that this does not introduce a memory leak, explosions are cleared
after each round.

Now with this warning fixed we can look into reintroducing the threaded
renderer, but it might still be unstable (and even related to these
thread shenanigans).