Port item to QSGImageNode
With this convenience class we avoid having to set geometry vertices
ourself.

But note that it still crashes with the threaded renderer with the very
same stacktrace:

QSGBatchRenderer::Renderer::renderUnmergedBatch (this=this@entry=0x7fffacd4a1d0,
    renderBatch=<optimized out>, depthPostPass=depthPostPass@entry=false)
    at /usr/src/debug/qt6-declarative/qtdeclarative-everywhere-src-6.6.1/src/quick/scenegraph/coreapi/qsgbatchrenderer.cpp:3418
3418            QSGGeometry *g = e->node->geometry();

There is a very similar Qt forum thread [0], where porting away from
QSGGeometryNode fixed the problem. Unfortunately we are not as lucky.

[0] https://forum.qt.io/post/604052